backend/ai_battle_api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "AI Battle Arena Backend"
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard", "testing"], version = "^0.115.7"}
psycopg = {extras = ["binary"], version = "^3.2.4"}
aiohttp = "^3.11.11"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
tweepy = "^4.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
